refactor(adapters): isolate xAI schema analysis (split S05 L1/3)
